+++ user/urls.py
@@ -2,9 +2,10 @@
 author: Sanidhya Mangal
 github: sanidhyamangal
 """
-from django.urls import path, include
-from user import views
+from django.urls import include, path
 from rest_framework.routers import DefaultRouter
+
+from user import views
 
 router = DefaultRouter()