+++ user/views.py
@@ -2,19 +2,20 @@
 author: Sanidhya Mangal
 github: sanidhyamangal
 """
-from rest_framework.decorators import action
-from rest_framework.response import Response
-from rest_framework.validators import ValidationError
-from rest_framework.permissions import IsAdminUser
-from rest_framework import status
-from rest_framework_simplejwt.tokens import RefreshToken
+from base.exceptions import BaseValidationError
 from base.viewsets import BaseAPIViewSet
+from rest_framework.decorators import action
+from rest_framework.permissions import AllowAny, IsAdminUser
+from rest_framework.response import Response
+from rest_framework_simplejwt.tokens import RefreshToken
+
+from .mixins import ForgotPasswordMixin
+from .models import User, VerifyUser
 from .permissions import AllowAnyPostReadUpdateDestroyOwnerOrAdmin
 from .serializers import UserSerializer
-from .models import User
 
 
-class UserViewSet(BaseAPIViewSet):
+class UserViewSet(BaseAPIViewSet, ForgotPasswordMixin):
     model_class = User
     serializer_class = UserSerializer
     instance_name = "user"
@@ -23,6 +24,8 @@ class UserViewSet(BaseAPIViewSet):
     def get_permissions(self):
         if self.action == "list":
             return [IsAdminUser()]
+        if self.action == "verify_user":
+            return [AllowAny()]
         return [AllowAnyPostReadUpdateDestroyOwnerOrAdmin()]
 
     @action(methods=["POST"], detail=False)
@@ -30,27 +33,18 @@ class UserViewSet(BaseAPIViewSet):
         try:
             for field in ['username', 'password']:
                 if not self.request.data.get(field):
-                    raise ValidationError({
-                        'status': False,
-                        'message': f"{field} is required",
-                        "data": {}
-                    })
+                    raise BaseValidationError(detail=f"{field} is required")
 
             _user = self.model_class.objects.get(
                 username=self.request.data['username'])
 
             if not _user.check_password(self.request.data['password']):
-                raise ValidationError({
-                    'status': False,
-                    'message': 'Incorrect password',
-                    'data': {}
-                })
+                raise BaseValidationError(detail=f"Incorrect Password")
 
             token = RefreshToken.for_user(_user)
             user_serializer = UserSerializer(_user)
             _data = user_serializer.data
             _data.update({'token': str(token.access_token)})
-
             return Response({
                 'status': True,
                 'message': 'Login Successful',
@@ -58,8 +52,24 @@ class UserViewSet(BaseAPIViewSet):
             })
 
         except self.model_class.DoesNotExist:
-            raise ValidationError({
-                'status': False,
-                'message': "User Doesn't Exist",
-                "data": {}
+            raise BaseValidationError(detail=f"User Doesn't Exists")
+
+    @action(methods=["GET"],
+            detail=False,
+            url_path="verify-user/(?P<token>[^/.]+)",
+            url_name="verify-user")
+    def verify_user(self, request, *args, **kwargs):
+        try:
+            token = kwargs.pop('token')
+            obj = VerifyUser.objects.get(pk=token)
+            if obj.is_used:
+                raise BaseValidationError(detail=f"Token already used")
+            obj.is_used = True
+            obj.save()
+            return Response({
+                'status': True,
+                'message': 'User validate successfully',
+                'data': {}
             })
+        except VerifyUser.DoesNotExist:
+            raise BaseValidationError(detail=f"Not a valid token")