+++ changelog.md
@@ -3,3 +3,10 @@
 - added: login ops for user
 - added: base skeleton for the project
 - added: simple JWT auth
+
+### 2020-10-31
+- added: Forgot password mixin
+- added: Base exception for the rest apis
+- added: forgot password and rest forgot password serializer
+- added: forgot password and verify user models
+- added: verify user functionality